+++ apps/docs/src/examples/extras/float/Scene.svelte
@@ -17,10 +17,7 @@
   interactivity()
 </script>
 
-<Environment
-  path="/hdr/"
-  files="shanghai_riverside_1k.hdr"
-/>
+<Environment url="/textures/equirectangular/hdr/shanghai_riverside_1k.hdr" />
 
 <Float
   rotationIntensity={0.15}